Revert "Camilla feature"
