--- test_package/scripts/talker_camilla.py
@@ -1,20 +1,0 @@
-#!/usr/bin/env python
-# license removed for brevity
-import rospy
-from std_msgs.msg import String
-
-def talker():
-    pub = rospy.Publisher('Camilla', String, queue_size=10)
-    rospy.init_node('talker_camilla', anonymous=True)
-    rate = rospy.Rate(2) # 2hz
-    while not rospy.is_shutdown():
-        hello_str = "Camilla %s" % rospy.get_time()
-        rospy.loginfo(hello_str)
-        pub.publish(hello_str)
-        rate.sleep()
-
-if __name__ == '__main__':
-    try:
-        talker()
-    except rospy.ROSInterruptException:
-        pass
